Adjustments:  - Added BehaviorSubject update variable to deal with backend status
Feature:
 - delete method

+++ src/app/events/events-list/events-list.component.ts
@@ -2,7 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {EventsService} from '../events.service';
 import {ModalService} from '../../shared/modal.service';
 import {Event} from '../event';
-import {delay, distinctUntilChanged, switchMap, take} from 'rxjs/operators';
+import {distinctUntilChanged, take} from 'rxjs/operators';
 import {formatCurrency, formatDate} from '@angular/common';
 import {EMPTY, Observable} from 'rxjs';
 
@@ -16,10 +16,12 @@ export class EventsListComponent implements OnInit {
   events$: Observable<Event[]>;
 
   constructor(private eventsService: EventsService,
-              private modalService: ModalService) { }
+              private modalService: ModalService
+             ) { }
 
   ngOnInit() {
     this.refresh();
+    this.eventsService.updated.subscribe(update => update ? this.refresh() : EMPTY);
   }
 
   refresh() {
@@ -31,15 +33,8 @@ export class EventsListComponent implements OnInit {
     this.openModal(event);
     this.modalService.confirm
       .pipe(take(1), distinctUntilChanged())
-      .subscribe(
-        confirm => {
-          if (confirm) {
-            this.eventsService.delete(event.id);
-            delay(1000);
-            this.refresh();
-          }
-        },
-        error => console.error(error)
+      .subscribe(confirm => confirm ? this.eventsService.delete(event.id) : EMPTY,
+                 error => console.error('erro ao excluir: ' + error)
       );
   }
